requirements.txt
# Requirements for Ejeh Ankpa Palace Platform
# Django Web Framework
Django>=4.2,<5.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Media Storage - Cloudinary
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0

# Static Files
whitenoise>=6.6.0

# Forms
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10

# Environment Variables
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.2.0

# Security
django-cors-headers>=4.3.1

# Utilities
Pillow>=10.1.0

# For Vercel deployment
setuptools>=69.0.0
